Incremental analyzis API
This paves the way for Rust support of streamed or incremental APIs, and
more specifically for avoiding DoS based on resource usage.
